Added ProblemDetails. refactored packages/classes
