+++ src/main/java/com/shyamalmadura/kafka/example/application/dto/User.java
@@ -1,9 +1,8 @@
-package com.shyamalmadura.kafka.example.dto;
+package com.shyamalmadura.kafka.example.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Data
