+++ src/main/java/com/shyamalmadura/kafka/example/controller/aop/RestExceptionHandlerAdvice.java
@@ -1,0 +1,20 @@
+package com.shyamalmadura.kafka.example.controller.aop;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ProblemDetail;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class RestExceptionHandlerAdvice {
+
+    @ExceptionHandler(Exception.class)
+    public ProblemDetail handleGenericError(Exception e){
+        return ProblemDetail.forStatusAndDetail(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ProblemDetail handle(IllegalArgumentException e){
+        return ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, e.getMessage());
+    }
+}
